Restructure all android files
