+++ android/app/src/androidTest/java/com/cloudsports/actiondetect/actiondetect/ExampleInstrumentedTest.kt
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.google.mediapipe.examples.poselandmarker", appContext.packageName)
+        assertEquals("com.cloudsports.actiondetect", appContext.packageName)
     }
 }
