+++ android/app/src/main/java/com/cloudsports/actiondetect/ChooseActionActivity.kt
@@ -1,0 +1,17 @@
+package com.cloudsports.actiondetect
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.cloudsports.actiondetect.item.Action
+
+
+class ChooseActionActivity : AppCompatActivity() {
+
+}
