+++ android/app/src/main/java/com/cloudsports/actiondetect/PoseLandmarkerHelper.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cloudsports.actiondetect.actiondetect
+package com.cloudsports.actiondetect
 
 import android.content.Context
 import android.graphics.Bitmap
