+++ android/app/src/main/java/com/cloudsports/actiondetect/algorithm/PoseSampleOutlier.kt
@@ -1,4 +1,6 @@
-package com.cloudsports.actiondetect.actiondetect.algorithm
+package com.cloudsports.actiondetect.algorithm
+
+import com.cloudsports.actiondetect.algorithm.PoseSample
 
 class PoseSampleOutlier(sample: PoseSample, detected_class: Map<String, Int>, all_classes: Map<String, Int>) {
     private val _sample = sample
