+++ android/app/src/main/java/com/cloudsports/actiondetect/algorithm/RepetitionCounter.kt
@@ -1,4 +1,4 @@
-package com.cloudsports.actiondetect.actiondetect.algorithm
+package com.cloudsports.actiondetect.algorithm
 
 import android.os.Build
 import androidx.annotation.RequiresApi
