+++ android/app/src/main/java/com/cloudsports/actiondetect/algorithm/Util/ArrayOperation.kt
@@ -1,4 +1,4 @@
-package com.cloudsports.actiondetect.actiondetect.algorithm.Util
+package com.cloudsports.actiondetect.algorithm.Util
 
 import kotlin.math.abs
 import kotlin.math.pow
