+++ android/app/src/main/java/com/cloudsports/actiondetect/algorithm/Util/MyDoubleArray.kt
@@ -1,4 +1,4 @@
-package com.cloudsports.actiondetect.actiondetect.algorithm.Util
+package com.cloudsports.actiondetect.algorithm.Util
 
 class MyDoubleArray(var array: DoubleArray) {
