+++ android/app/src/main/java/com/cloudsports/actiondetect/debug/ToastDebug.kt
@@ -1,4 +1,4 @@
-package com.cloudsports.actiondetect.actiondetect.debug
+package com.cloudsports.actiondetect.debug
 
 import android.content.Context;
 import android.os.Handler;
