+++ android/app/src/main/java/com/cloudsports/actiondetect/fragment/GalleryFragment.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cloudsports.actiondetect.actiondetect.fragment
+package com.cloudsports.actiondetect.fragment
 
 import android.graphics.Bitmap
 import android.graphics.ImageDecoder
@@ -32,9 +32,9 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import com.cloudsports.actiondetect.actiondetect.MainViewModel
-import com.cloudsports.actiondetect.actiondetect.PoseLandmarkerHelper
-import com.google.mediapipe.examples.actiondetect.databinding.FragmentGalleryBinding
+import com.cloudsports.actiondetect.MainViewModel
+import com.cloudsports.actiondetect.PoseLandmarkerHelper
+import com.cloudsports.actiondetect.databinding.FragmentGalleryBinding
 import com.google.mediapipe.tasks.vision.core.RunningMode
 import java.util.*
 import java.util.concurrent.Executors
