+++ android/app/src/main/java/com/cloudsports/actiondetect/fragment/PermissionsFragment.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cloudsports.actiondetect.actiondetect.fragment
+package com.cloudsports.actiondetect.fragment
 
 import android.Manifest
 import android.content.Context
@@ -25,7 +25,7 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.Navigation
-import com.google.mediapipe.examples.actiondetect.R
+import com.cloudsports.actiondetect.R
 
 private val PERMISSIONS_REQUIRED = arrayOf(Manifest.permission.CAMERA)
