+++ android/app/src/main/java/com/cloudsports/actiondetect/item/Action.kt
@@ -1,0 +1,7 @@
+package com.cloudsports.actiondetect.item
+
+class Action(var title: String, var msg: String, var image: String, var localImage: String) {
+    override fun toString(): String {
+        return "Action(title='$title', msg='$msg', image='$image', localImage='$localImage')"
+    }
+}
