+++ android/app/src/main/res/layout/fragment_camera.xml
@@ -26,7 +26,7 @@
         android:layout_height="match_parent"
         app:scaleType="fillStart" />
 
-    <com.cloudsports.actiondetect.actiondetect.OverlayView
+    <com.cloudsports.actiondetect.OverlayView
         android:id="@+id/overlay"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
