+++ android/app/src/main/res/layout/fragment_gallery.xml
@@ -51,7 +51,7 @@
                 android:layout_height="match_parent"
                 android:visibility="gone" />
 
-            <com.cloudsports.actiondetect.actiondetect.OverlayView
+            <com.cloudsports.actiondetect.OverlayView
                 android:id="@+id/overlay"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
